requirement.txt
langchain-core>=1.2.0
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langgraph==1.0.4
langsmith==0.4.59
pydantic>=2.0.0
streamlit>=1.52.0